clean opensolar, define a template for platforms to follow
standardize naming of some parts so that we can have other platforms
modelled in the same way
