+++ platforms/opensolar/auctions.go
@@ -7,7 +7,7 @@ import (
 // Contract auctions are specific for public infrastructure and for projects with multiple stakeholders
 // where transparency is important. The main point is to avoid corruption and produce positive competition
 // among providers. When funding solar project using international investor money, transparency and auditability
-// to them is  also a crucial aspect since they are often not close to the project, and want to make sure there
+// to them is also a crucial aspect since they are often not close to the project, and want to make sure there
 // is a tobust due diligence prior to unlocking the funds.
 
 // Different Auctions or Tenders are designed based on the nature of the project.
@@ -19,19 +19,19 @@ import (
 // the school would want the most stuff but you need to vet which contracts are good
 // and not.
 
+// For more trivia, see https://en.wikipedia.org/wiki/Auction
 type ContractAuction struct {
 	// TODO: this struct isn't used yet as it needs handlers and stuff, but when
 	// we move off main.go for testing, this must be used in order to make stuff
 	// easier for us.
 	AllContracts    []Project
 	AllContractors  []Entity
-	WinningContract Project // do we need this?
+	WinningContract Project
 }
 
-// auctions contains stuff related to choosing the best contract and potentially
-// future auction logic that might need to be housed here
+// SelectContractBlind selects the winning bid  based on blind auctio nrules
+// in a blind auction, the bid with the highest price wins
 func SelectContractBlind(arr []Project) (Project, error) {
-	// in a blind auction, the bid with the highest price wins
 	var a Project
 	if len(arr) == 0 {
 		return a, fmt.Errorf("Empty array passed!")
@@ -39,7 +39,7 @@ func SelectContractBlind(arr []Project) (Project, error) {
 	// array is not empty, min 1 elem
 	a = arr[0]
 	for _, elem := range arr {
-		if elem.Params.TotalValue < a.Params.TotalValue {
+		if elem.TotalValue < a.TotalValue {
 			a = elem
 			continue
 		}
@@ -47,6 +47,8 @@ func SelectContractBlind(arr []Project) (Project, error) {
 	return a, nil
 }
 
+// SelectContractVickrey selects the winning bid based on vickrey auction rules
+// in a vickrey auction, the bid with the second highest price wins
 func SelectContractVickrey(arr []Project) (Project, error) {
 	var winningContract Project
 	if len(arr) == 0 {
@@ -56,7 +58,7 @@ func SelectContractVickrey(arr []Project) (Project, error) {
 	winningContract = arr[0]
 	var pos int
 	for i, elem := range arr {
-		if elem.Params.TotalValue < winningContract.Params.TotalValue {
+		if elem.TotalValue < winningContract.TotalValue {
 			winningContract = elem
 			pos = i
 			continue
@@ -70,27 +72,28 @@ func SelectContractVickrey(arr []Project) (Project, error) {
 		// means only one contract was proposed for this project, so fall back to blind auction
 		return winningContract, nil
 	}
-	vickreyPrice := arr[0].Params.TotalValue
+	vickreyPrice := arr[0].TotalValue
 	for _, elem := range arr {
-		if elem.Params.TotalValue < vickreyPrice {
-			vickreyPrice = elem.Params.TotalValue
+		if elem.TotalValue < vickreyPrice {
+			vickreyPrice = elem.TotalValue
 		}
 	}
 	// we have the winner, who's elem and we have the price which is vickreyPrice
-	// voerwrite the winning contractor's contract
-	winningContract.Params.TotalValue = vickreyPrice
+	// overwrite the winning contractor's contract
+	winningContract.TotalValue = vickreyPrice
 	return winningContract, winningContract.Save()
 }
 
+// SelectContractTime selects the winning contract based on the least time to completion
 func SelectContractTime(arr []Project) (Project, error) {
 	var a Project
 	if len(arr) == 0 {
 		return a, fmt.Errorf("Empty array passed!")
 	}
-	// array is not empty, min 1 elem
+
 	a = arr[0]
 	for _, elem := range arr {
-		if elem.Params.Years < a.Params.Years {
+		if elem.Years < a.Years {
 			a = elem
 			continue
 		}
@@ -98,8 +101,8 @@ func SelectContractTime(arr []Project) (Project, error) {
 	return a, nil
 }
 
+// SetAuctionType sets the auction type of a specific project
 func (project *Project) SetAuctionType(auctionType string) error {
-	// see https://en.wikipedia.org/wiki/Auction for primary auction types
 	switch auctionType {
 	case "blind":
 		project.AuctionType = "blind"
