+++ platforms/opensolar/entity_developer.go
