+++ platforms/opensolar/entity_guarantor.go
