+++ platforms/template/auctions.go
@@ -1,0 +1,4 @@
+package template
+
+// contains various handler functions which perform the function of choosing between a
+// set of contracts and selecting the winning one
