+++ platforms/template/contract.go
@@ -1,0 +1,9 @@
+package template
+
+// the contract that powers the specific platform
+// the contract must at minimum have five functions similar in working to the examples below:
+// 1. PreInvestmentCheck - a function which cheks if a specific investor can invest in a particular project
+// 2. Invest - the main investment function which handles generation of receipt assets and similar
+// 3. UpdateProjectAfterInvestment - Update the project after receipt of investor assets by the investor
+// 4. sendRecipientAssets - send the recipient(s) asset(s) on acceptance of investment
+// 5. UpdateProjectAfterAcceptance - update the project structure after acceptance by recipient
