+++ platforms/template/entities.go
@@ -1,0 +1,8 @@
+package template
+
+// contains the different entities involved in the platform apart from the base entities like the investor and the recipient
+// and example of the different entities involved in the opensolar project are:
+// 1. Originator
+// 2. Contractor
+// 3.  Guarantor
+// 4. Developer
