+++ platforms/template/entity_2.go
@@ -1,0 +1,3 @@
+package template
+
+// contains the functions and handlers related to entity 2
