+++ platforms/template/feedback.go
@@ -1,0 +1,4 @@
+package template
+
+// contains the feedback strcucture which can be used to give feedback on various entities involved in a contract / project
+// this feedback structure would also define the reputation models that is put in place
