+++ platforms/template/helpers.go
@@ -1,0 +1,3 @@
+package template
+
+// contains helper functions not belonging to a specific category
