+++ platforms/template/template.go
@@ -1,0 +1,6 @@
+package template
+
+// contains the definition of the project structure
+// it is necessary to have atleast one structure (eg the Project structure in opensolar)
+// which gives a high level overview of what the investors and investing in, how the
+// recipient is defined, how the other entities interact with it, etc.
